Change date and time pickers on meetings(aka monthlies) to match those on workshops
