+++ app/models/meeting.rb
@@ -1,6 +1,9 @@
 class Meeting < ActiveRecord::Base
+  include DateTimeConcerns
   include Listable
   include Invitable
+
+  attr_accessor :local_date, :local_time
 
   resourcify :permissions, role_cname: 'Permission', role_table_name: :permission
 
@@ -10,8 +13,10 @@ class Meeting < ActiveRecord::Base
   has_and_belongs_to_many :chapters
 
   validates :date_and_time, :venue, presence: true
+  validates :date_and_time, presence: true, if: proc { |model| model.venue_id.present? }
   validates :slug, uniqueness: true, if: proc { |model| model.slug.present? }
 
+  before_validation :set_date_and_time
   before_save :set_slug
 
   def invitees
@@ -26,14 +31,6 @@ class Meeting < ActiveRecord::Base
     slug
   end
 
-  def date
-    I18n.l(date_and_time, format: :dashboard)
-  end
-
-  def past?
-    date_and_time < Time.zone.today
-  end
-
   def attending?(member)
     invitations.accepted.where(member: member).present?
   end
@@ -46,7 +43,15 @@ class Meeting < ActiveRecord::Base
     CSV.generate { |csv| attendees_array.each { |a| csv << a } }
   end
 
+  def time_zone
+    'London'
+  end
+
   private
+
+  def attendees_array
+    invitations.accepted.map { |a| [a.member.full_name] }
+  end
 
   def set_slug
     return if slug.present?
@@ -54,9 +59,5 @@ class Meeting < ActiveRecord::Base
       url = "#{I18n.l(date_and_time, format: :year_month).downcase}-#{title.parameterize}-#{index+1}"
       break url unless Meeting.where(slug: url).exists?
     end
-  end
-
-  def attendees_array
-    invitations.accepted.map { |a| [a.member.full_name] }
   end
 end
