+++ app/models/workshop.rb
@@ -1,4 +1,5 @@
 class Workshop < ActiveRecord::Base
+  include DateTimeConcerns
   include Invitable
   include Listable
 
@@ -52,17 +53,9 @@ class Workshop < ActiveRecord::Base
     has_host? && host.address.present?
   end
 
-  def past?
-    date_and_time.past?
-  end
-
   def rsvp_available?
     return rsvp_closes_at.future? if rsvp_closes_at
     future?
-  end
-
-  def future?
-    date_and_time.future?
   end
 
   def open_for_rsvp?
@@ -103,40 +96,15 @@ class Workshop < ActiveRecord::Base
     WorkshopPolicy
   end
 
-  def date_and_time
-    return nil unless super
-    super.in_time_zone(time_zone)
-  end
-
   def rsvp_opens_at
     return nil unless super
     super.in_time_zone(time_zone)
   end
 
-  def date
-    I18n.l(date_and_time, format: :dashboard)
-  end
-
-  def time
-    date_and_time.try(:time)
-  end
-
   private
-
-  def set_date_and_time
-    new_date_and_time = datetime_from_fields(local_date, local_time)
-    self.date_and_time = new_date_and_time if new_date_and_time
-  end
 
   def set_opens_at
     new_opens_at = datetime_from_fields(rsvp_open_local_date, rsvp_open_local_time)
     self.rsvp_opens_at = new_opens_at if new_opens_at
-  end
-
-  def datetime_from_fields(date_string, time_string)
-    return nil if date_string.blank? || time_string.blank? || !time_zone
-    date = Date.parse(date_string)
-    time = Time.parse(time_string)
-    ActiveSupport::TimeZone[time_zone].local(date.year, date.month, date.day, time.hour, time.min)
   end
 end
