+++ app/views/admin/meetings/_form.html.haml
@@ -7,8 +7,11 @@
         .medium-6.columns
           = f.input :slug
   .row
-    .large-12.columns
-      = f.input :date_and_time, as: :datetime, required: true, label: 'Date', input_html: { style: 'width: 120px;' }
+    .medium-6.columns.admin_margin
+      = f.input :local_date, as: :string, required: true, input_html: { data: { value: @meeting.date_and_time.try(:strftime, '%d/%m/%Y') } }
+  .row
+    .medium-6.columns
+      = f.input :local_time, as: :string, required: true, input_html: { data: { value: @meeting.time.try(:strftime, '%H:%M') } }
   .row
     .large-12.columns.admin_margin
       = f.input :description, placeholder: 'Supports HTML', as: :text, input_html: { style: 'height: 300px;' }
